MOVEONLY: Consensus: Make IsSuperMajority static and move to blockverify.cpp:
from main:
-CheckBlock
-ContextualCheckBlock
-GetBlockSubsidy
